# app/Http/Controllers/AnimeController.php
<?php

namespace App\Http\Controllers;

use App\Anime;
use App\AnimeRating;
use App\Events\AnimeViewed;
use App\Helpers\JSONResult;
use App\Http\Resources\ActorCharacterAnimeResource;
use App\Http\Resources\ActorResource;
use App\Http\Resources\AnimeRelationsResource;
use App\Http\Resources\AnimeResource;
use App\Http\Resources\AnimeResourceBasic;
use App\Http\Resources\AnimeSeasonResource;
use App\Http\Resources\CharacterResource;
use Exception;
use Illuminate\Http\JsonResponse;
use Illuminate\Support\Facades\Auth;
use Illuminate\Support\Facades\Validator;
use Illuminate\Http\Request;

class AnimeController extends Controller
{
    /**
     * Returns detailed information about an Anime.
     *
     * @param Anime $anime
     * @return JsonResponse
     */
    public function view(Anime $anime): JsonResponse
    {
        // Call the AnimeViewed event
        AnimeViewed::dispatch($anime);

        // Show the Anime details response
        return JSONResult::success([
            'data' => AnimeResource::collection([$anime])
        ]);
    }

    /**
     * Returns actor information about an Anime.
     *
     * @param Anime $anime
     * @return JsonResponse
     */
    public function actorsAnime(Anime $anime): JsonResponse
    {
        // Get the actors
        $actors = $anime->getActors();

        return JSONResult::success([
            'data' => ActorResource::collection($actors)
        ]);
    }

    /**
     * Returns character information about an Anime.
     *
     * @param Anime $anime
     * @return JsonResponse
     */
    public function charactersAnime(Anime $anime): JsonResponse
    {
        // Get the actors
        $actors = $anime->getCharacters();

        return JSONResult::success([
            'data' => CharacterResource::collection($actors)
        ]);
    }

    /**
     * Returns actor-character-anime information about an Anime.
     *
     * @param Anime $anime
     * @return JsonResponse
     */
    public function actorCharacterAnime(Anime $anime): JsonResponse
    {
        // Get the actor-character-anime
        $actorCharacterAnime = $anime->getActorCharacterAnime();

        return JSONResult::success([
            'data' => ActorCharacterAnimeResource::collection($actorCharacterAnime)
        ]);
    }

    /**
     * Returns relations information about an Anime.
     *
     * @param Anime $anime
     * @return JsonResponse
     */
    public function relationsAnime(Anime $anime): JsonResponse
    {
        // Get the actors
        $relations = $anime->getAnimeRelations();

        return JSONResult::success([
            'data' => AnimeRelationsResource::collection($relations)
        ]);
    }

    /**
     * Returns season information for an Anime
     *
     * @param Anime $anime
     * @return JsonResponse
     */
    public function seasonsAnime(Anime $anime): JsonResponse
    {
        // Get the seasons
        $seasons = $anime->getSeasons();

        return JSONResult::success([
            'data' => AnimeSeasonResource::collection($seasons)
        ]);
    }

    /**
     * Adds a rating for an Anime item
     *
     * @param Request $request
     * @param Anime   $anime
     * @return JsonResponse
     * @throws Exception
     */
    public function rateAnime(Request $request, Anime $anime): JsonResponse
    {
        if (!Auth::user()->isTracking($anime))
            return JSONResult::error('Please add ' . $anime->title . ' to your library first.');

        // Validate the inputs
        $validator = Validator::make($request->all(), [
            'rating' => 'bail|required|numeric|between:' . AnimeRating::MIN_RATING_VALUE . ',' . AnimeRating::MAX_RATING_VALUE
        ]);

        // Check validator
        if ($validator->fails())
            return JSONResult::error($validator->errors()->first());

        // Fetch the variables
        $givenRating = $request->input('rating');

        // Try to modify the rating if it already exists
        $foundRating = AnimeRating::where([
            ['anime_id', '=', $anime->id],
            ['user_id', '=', Auth::id()]
        ])->first();

        // The rating exists
        if($foundRating) {
            // If the given rating is 0, delete the rating
            if($givenRating <= 0)
                $foundRating->delete();
            // Update the current rating
            else {
                $foundRating->rating = $givenRating;
                $foundRating->save();
            }
        }
        // Rating needs to be inserted
        else {
            // Only insert the rating if it's rated higher than 0
            if($givenRating > 0) {
                AnimeRating::create([
                    'anime_id'  => $anime->id,
                    'user_id'   => Auth::id(),
                    'rating'    => $givenRating
                ]);
            }
        }

        return JSONResult::success();
    }

    /**
     * Retrieves Anime search results
     *
     * @param Request $request
     * @return JsonResponse
     */
    public function search(Request $request): JsonResponse
    {
        // Validate the inputs
        $validator = Validator::make($request->all(), [
            'query' => 'bail|required|string|min:1'
        ]);

        // Check validator
        if($validator->fails())
            return JSONResult::error($validator->errors()->first());

        $searchQuery = $request->input('query');

        // Search for the Anime
        $resultArr = Anime::kuroSearch($searchQuery, [
            'limit' => Anime::MAX_SEARCH_RESULTS
        ]);

        return JSONResult::success([
            'max_search_results'    => Anime::MAX_SEARCH_RESULTS,
            'data'                  => AnimeResourceBasic::collection($resultArr)
        ]);
    }
}