# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "This is the documentation for Kurozora API.",
    "version": "1.0.8",
    "title": "Kurozora API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "kurozoraapp@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "https://kurozora.app/api/v1",
      "description": "Kurozora LIVE"
    },
    {
      "url": "http://kurozora-web.test/api/v1",
      "description": "Kurozora LOCAL"
    }
  ],
...
  "paths": {
...
    "/actors": {
      "get": {
        "tags": [
          "actors"
        ],
        "operationId": "getActors",
        "summary": "Actors list.",
        "description": "This endpoint will retrieve the actors.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSONSuccess"
          },
          "400": {
            "$ref": "#/components/responses/JSONError"
          }
        }
      }
    },
...
    "/anime/{animeID}/relations": {
      "get": {
        "security": [
          {
            "kurozoraBearer": [
            ]
          }
        ],
        "tags": [
          "anime"
        ],
        "operationId": "getAnimeRelations",
        "summary": "Anime all relations information.",
        "description": "This endpoint will retrieve the relations of an Anime item.",
        "parameters": [
          {
            "in": "path",
            "name": "animeID",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID of the anime."
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GetAnimeRelations"
          },
          "400": {
            "$ref": "#/components/responses/JSONError"
          }
        }
      }
    },
...
    "/characters": {
      "get": {
        "tags": [
          "characters"
        ],
        "operationId": "getCharacters",
        "summary": "Characters list.",
        "description": "This endpoint will retrieve the characters.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/GetCharacters"
          },
          "400": {
            "$ref": "#/components/responses/JSONError"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "kurozoraBearer": {
        "type": "apiKey",
        "in": "header",
        "name": "kuro-auth"
      }
    },
    "responses": {
      "JSONSuccess": {
        "description": "Successful Kurozora JSON response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "meta": {
                  "version": "1.0.8",
                  "query_count": 0,
                  "is_user_authenticated": false,
                  "authenticated_user_id": null
                }
              }
            }
          }
        }
      },
      "JSONError": {
        "description": "Error Kurozora JSON response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "meta": {
                  "version": "1.0.8",
                  "query_count": 0,
                  "is_user_authenticated": false,
                  "authenticated_user_id": null
                },
                "message": "Description of the error."
              }
            }
          }
        }
      },
...
      "GetAnimeRelations": {
        "description": "Successful GET Anime Relations JSON response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "data": [
                  {
                    "show": {
                      "id": 217,
                      "type": "show",
                      "href": "/api/v1/anime/217",
                      "attributes": {
                        "title": "Omnis odit hic dolorem itaque.",
                        "type": "Special",
                        "source": "4-Koma Manga",
                        "anidb_id": null,
                        "anilist_id": null,
                        "kitsu_id": null,
                        "imdb_id": null,
                        "mal_id": null,
                        "network": null,
                        "studio": [],
                        "status": "Tba",
                        "episodes": 0,
                        "seasons": 0,
                        "average_rating": 0,
                        "rating_count": 0,
                        "synopsis": "Aspernatur aperiam corrupti similique voluptatibus qui. Voluptas eum ea doloribus qui inventore hic. Voluptas accusantium libero autem ipsum molestiae aut nobis. Aliquam unde voluptates corrupti ut temporibus eum.",
                        "runtime": 25,
                        "watch_rating": "TV-Y7 (FV)",
                        "tagline": "Sunt error hic quod sint.",
                        "video_url": null,
                        "poster": null,
                        "background": null,
                        "nsfw": true,
                        "genres": [],
                        "first_aired": "2014-12-10",
                        "last_aired": "2015-05-27",
                        "air_time": "16:49:00",
                        "air_day": 7,
                        "copyright": "℗ 1997 Kulas-Baumbach"
                      },
                      "current_user": {
                        "given_rating": 0,
                        "library_status": null,
                        "is_favorite": false
                      }
                    },
                    "attributes": {
                      "type": "Full story"
                    }
                  },
                  {
                    "show": {
                      "id": 219,
                      "type": "show",
                      "href": "/api/v1/anime/219",
                      "attributes": {
                        "title": "Nam voluptatum voluptas optio qui.",
                        "type": "Unknown",
                        "source": "Original",
                        "anidb_id": null,
                        "anilist_id": null,
                        "kitsu_id": null,
                        "imdb_id": null,
                        "mal_id": null,
                        "network": null,
                        "studio": [],
                        "status": "Tba",
                        "episodes": 0,
                        "seasons": 0,
                        "average_rating": 0,
                        "rating_count": 0,
                        "synopsis": "Rerum distinctio quibusdam porro et. Modi natus iusto ut qui praesentium dicta sed ratione. Porro excepturi et qui laudantium tenetur non iure.",
                        "runtime": 25,
                        "watch_rating": "TV-Y7",
                        "tagline": "Expedita esse et cupiditate id impedit qui ut.",
                        "video_url": null,
                        "poster": null,
                        "background": null,
                        "nsfw": false,
                        "genres": [],
                        "first_aired": "2015-02-10",
                        "last_aired": "2015-05-05",
                        "air_time": "18:32:00",
                        "air_day": 6,
                        "copyright": "® 1976 Swift Ltd"
                      },
                      "current_user": {
                        "given_rating": 0,
                        "library_status": null,
                        "is_favorite": false
                      }
                    },
                    "attributes": {
                      "type": "Adaptation"
                    }
                  }
                ]
              }
            }
          }
        }
      },
...
      "GetCharacters": {
        "description": "Successful GET Characters JSON response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "data": [
                  {
                    "id": 1,
                    "type": "characters",
                    "href": "/api/v1/characters/1",
                    "attributes": {
                      "name": "Tyra Armstrong IV",
                      "about": "Aspernatur odio ab laborum neque voluptatem quia amet aliquam. Quaerat autem quo quis reprehenderit iure nostrum. Explicabo nam dignissimos ut consequatur. Veniam sapiente explicabo laborum sit harum optio enim sed. Ex corrupti fuga esse quia voluptatibus. Sit in qui ut saepe adipisci reiciendis. Consequatur aspernatur molestiae est saepe mollitia qui. Laudantium eligendi odio adipisci illo. Culpa delectus maxime aperiam nemo impedit eos. Dicta et aliquid ratione est alias autem possimus. Aut repellendus beatae ipsum ipsum numquam animi. Nam et et est pariatur. Facilis sed et sit quia quam quo. A quae repudiandae pariatur mollitia. Exercitationem earum deserunt voluptate dolore facilis impedit fuga. Hic delectus consequatur quidem. Quia dolorum et voluptatum consectetur quis. Voluptatum hic in deleniti voluptates sunt. Qui perspiciatis rerum et porro.",
                      "image": "https://lorempixel.com/640/480/?46141",
                      "debut": "Episode 05",
                      "status": "Deceased",
                      "blood_type": "W",
                      "favorite_food": "numquam",
                      "bust": "27.70",
                      "waist": "78.44",
                      "hip": "53.99",
                      "height": "875",
                      "age": 265,
                      "birth_day": 1,
                      "birth_month": 7,
                      "astrological_sign": "Virgo ♍️"
                    }
                  },
                  {
                    "id": 2,
                    "type": "characters",
                    "href": "/api/v1/characters/2",
                    "attributes": {
                      "name": "Lorna Pollich",
                      "about": "Est vitae et voluptas ullam unde id ea. Quo non distinctio ipsa blanditiis aspernatur. Perferendis aut officia sit similique soluta odit eum. Sunt quo delectus est quis sapiente. A natus voluptatem omnis distinctio quis quas. Molestiae ut aut voluptas incidunt necessitatibus nihil culpa. Molestiae dolores illo dignissimos amet sit. Sit non et enim doloremque ratione exercitationem. Hic voluptas ab eos quia quia quia. Hic expedita est distinctio quaerat accusamus.",
                      "image": "https://lorempixel.com/640/480/?51170",
                      "debut": "Episode 82",
                      "status": "Deceased",
                      "blood_type": "A",
                      "favorite_food": "hic",
                      "bust": "22.62",
                      "waist": "33.46",
                      "hip": "37.65",
                      "height": "704",
                      "age": 292,
                      "birth_day": 1,
                      "birth_month": 2,
                      "astrological_sign": "Cancer ♋️"
                    }
                  }
                ]
              }
            }
          }
        }
      },
...
    }
  },
...
}